Moved everything outside of src/clm/python folder and deleted it
