+++ src/clm/commands/inner_sample_molecules_RNN.py
@@ -8,7 +8,7 @@ import torch
 from tqdm import tqdm
 
 from clm.datasets import Vocabulary, SelfiesDataset
-from clm.python.models import RNN
+from clm.models import RNN
 from clm.functions import set_seed, seed_type, read_smiles
